Refactor memgraph.cpp (#1164)
